Moved WCS tests to test_wcs.py and updated/fixed test for determining coord_meta from a WCS
